+++ services/updater.ts
@@ -13,7 +13,8 @@ import keys from "../utils/keys";
 import dumpProcessor from "./dumpProcessor";
 import termParser from "../scrapers/classes/parsersxe/termParser";
 import { Section as ScrapedSection } from "../types/types";
-import notifyer from "./notifyer";
+import { sendNotifications } from "./notifyer";
+import { NotificationInfo } from "../types/notifTypes";
 
 // ======= TYPES ======== //
 // A collection of structs for simpler querying of pre-scrape data
@@ -21,33 +22,6 @@ interface OldData {
   watchedSectionLookup: Record<string, Section>;
   watchedCourseLookup: Record<string, Course>;
   oldSectionsByClass: Record<string, string[]>;
-}
-
-// Stores information for all changes to a course or section
-export interface NotificationInfo {
-  updatedCourses: CourseNotificationInfo[];
-  updatedSections: SectionNotificationInfo[];
-}
-
-// marks new sections being added to a Course
-export interface CourseNotificationInfo {
-  subject: string;
-  courseId: string;
-  termId: string;
-  courseHash: string;
-  numberOfSectionsAdded: number;
-  campus: string;
-}
-
-// marks seats becoming available in a section
-export interface SectionNotificationInfo {
-  subject: string;
-  courseId: string;
-  sectionHash: string;
-  termId: string;
-  seatsRemaining: number;
-  crn: string;
-  campus: string;
 }
 
 // the types of models/records that a user can follow
@@ -115,7 +89,7 @@ class Updater {
 
     const totalTime = Date.now() - startTime;
 
-    await notifyer.sendNotifications(
+    await sendNotifications(
       notificationInfo,
       courseHashToUsers,
       sectionHashToUsers
